hammer_systems/requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==3.2.20
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
isort==5.11.5
packaging==23.1
pep8==1.7.1
phonenumbers==8.13.18
psycopg2-binary==2.8.6
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
gunicorn==20.1.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
